fast_django/pyproject.toml
[tool.poetry]
name = "fastdjango"
version = "0.0.1"
description = ""
authors = ["FastDjangoTeam"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.92.0"
Django = "^4.1.7"
psycopg2-binary = "^2.9.5"
pydantic = "^1.10.5"
redis = "^4.5.1"
django-environ = "^0.9.0"
uvicorn = "^0.20.0"
gunicorn = "^20.1.0"
whitenoise = "^6.3.0"
httpx = "^0.23.3"
async-lru = "^1.0.3"

[tool.poetry.group.dev.dependencies]
locust = "^2.14.2"
flake8 = "^6.0.0"
flake8-isort = "^6.0.0"
trio = "^0.22.0"
pytest-django = "^4.5.2"
asgi-lifespan = "^2.0.0"
nest-asyncio = "^1.5.6"
flake8-docstrings = "^1.7.0"
pytest-asyncio = "^0.20.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
